+++ boilingcore/boilingcore_test.go
@@ -148,3 +148,73 @@ func outputCompileErrors(buf *bytes.Buffer, outFolder string) {
 		fh.Close()
 	}
 }
+
+func TestProcessTypeReplacements(t *testing.T) {
+	s := new(State)
+	s.Config = &Config{}
+	s.Config.Imports.BasedOnType = make(map[string]importers.Set)
+	s.Tables = []drivers.Table{
+		{
+			Columns: []drivers.Column{
+				{
+					Name:     "id",
+					Type:     "int",
+					DBType:   "serial",
+					Default:  "some db nonsense",
+					Nullable: false,
+				},
+				{
+					Name:     "name",
+					Type:     "null.String",
+					DBType:   "serial",
+					Default:  "some db nonsense",
+					Nullable: true,
+				},
+			},
+		},
+	}
+
+	s.Config.TypeReplaces = []TypeReplace{
+		{
+			Match: drivers.Column{
+				DBType: "serial",
+			},
+			Replace: drivers.Column{
+				Type: "excellent.Type",
+			},
+			Imports: importers.Set{
+				ThirdParty: []string{`"rock.com/excellent"`},
+			},
+		},
+		{
+			Match: drivers.Column{
+				Type:     "null.String",
+				Nullable: true,
+			},
+			Replace: drivers.Column{
+				Type: "int",
+			},
+			Imports: importers.Set{
+				Standard: []string{`"context"`},
+			},
+		},
+	}
+
+	if err := s.processTypeReplacements(); err != nil {
+		t.Fatal(err)
+	}
+
+	if typ := s.Tables[0].Columns[0].Type; typ != "excellent.Type" {
+		t.Error("type was wrong:", typ)
+	}
+	if i := s.Config.Imports.BasedOnType["excellent.Type"].ThirdParty[0]; i != `"rock.com/excellent"` {
+		t.Error("imports were not adjusted")
+	}
+
+	if typ := s.Tables[0].Columns[1].Type; typ != "int" {
+		t.Error("type was wrong:", typ)
+	}
+	if i := s.Config.Imports.BasedOnType["int"].Standard[0]; i != `"context"` {
+		t.Error("imports were not adjusted")
+	}
+}